LayoutTests/media/encrypted-media/encrypted-media-constants-expected.txt

Test WebKitMediaKeyError constants and additional MediaError constant.

EXPECTED (WebKitMediaKeyError.MEDIA_KEYERR_UNKNOWN == '1') OK
EXPECTED (WebKitMediaKeyError.MEDIA_KEYERR_CLIENT == '2') OK
EXPECTED (WebKitMediaKeyError.MEDIA_KEYERR_SERVICE == '3') OK
EXPECTED (WebKitMediaKeyError.MEDIA_KEYERR_OUTPUT == '4') OK
EXPECTED (WebKitMediaKeyError.MEDIA_KEYERR_HARDWARECHANGE == '5') OK
EXPECTED (WebKitMediaKeyError.MEDIA_KEYERR_DOMAIN == '6') OK

EXPECTED (MediaError.MEDIA_ERR_ENCRYPTED == '5') OK
END OF TEST
